requirements.txt
mcp
PyGithub==2.1.1
streamlit==1.28.0
python-dotenv==1.0.0
aiohttp==3.9.0